refactor(cli): extract secrets cluster into commands/secrets.ts (5.0-alpha)
Second cut of the 5.0-alpha cli.ts god-module split. Moves the 10-handler
`secrets` command cluster (cmdSecrets + its 9 internally-routed sub-handlers:
validate, onecli, revoke-old, set, propagate, purge-history, migrate,
vault-migrate, sync, rotate) plus the private helper readSecretValueFromVault
out of cli.ts into src/commands/secrets.ts, matching the
scan.ts/security.ts pattern.

- cmdSecrets is the only export (the COMMANDS-table entry); the sub-handlers +
  readSecretValueFromVault stay module-private, reached via cmdSecrets'
  process.argv[3] routing.
- cli.ts imports cmdSecrets back; the COMMANDS dispatch table is unchanged.
- ensureSecretsBackend intentionally stays in cli.ts (cmdLogin uses it).
- Prunes now-unused secrets-only imports from cli.ts (secrets-validate,
  secrets-migrate, secrets-rotate-cli, secrets-set, secrets-purge-history,
  secrets-propagate, secrets-onecli, secrets-sync, secret-backends,
  requireElevation, SecretKeyConfig); keeps consumeElevation/vaultMeta/etc.
  still used elsewhere.
- Dynamic imports re-pathed ./ -> ../.

cli.ts: 5601 -> 4544 lines (-1057; -1786 cumulative across PRs 1-2).
Typecheck, eslint (0 errors), full test suite (2580), public-surface
snapshot, and prettier all green.
